# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_yamovo_contentx_internal_models.Permission:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      module:
        type: string
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_yamovo_contentx_internal_models.Role:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      is_system:
        type: boolean
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Permission'
        type: array
      slug:
        type: string
      updated_at:
        type: string
      user_count:
        type: integer
    required:
    - name
    type: object
...
  github_com_yamovo_contentx_internal_models.UserPreferences:
    properties:
      default_post_status:
        type: string
      email_notify:
        type: boolean
      items_per_page:
        type: integer
      language:
        type: string
      markdown_editor:
        type: boolean
      theme:
        type: string
    type: object
  github_com_yamovo_contentx_internal_models.UserStatus:
    enum:
    - active
    - inactive
    - banned
    - pending
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusInactive
    - UserStatusBanned
    - UserStatusPending
...
  github_com_yamovo_contentx_internal_services.CreateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permission_ids:
        items:
          type: integer
        type: array
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
  github_com_yamovo_contentx_internal_services.CreateUserRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
      role_id:
        type: integer
      status:
        type: string
      username:
        maxLength: 64
        minLength: 3
        type: string
    required:
    - email
    - password
    - role_id
    - username
    type: object
...
  github_com_yamovo_contentx_internal_services.SafeUser:
    properties:
      avatar:
        type: string
      bio:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      login_count:
        type: integer
      preferences:
        $ref: '#/definitions/github_com_yamovo_contentx_internal_models.UserPreferences'
      role:
        $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Role'
      status:
        $ref: '#/definitions/github_com_yamovo_contentx_internal_models.UserStatus'
      username:
        type: string
      website:
        type: string
    type: object
...
  github_com_yamovo_contentx_internal_services.UpdateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permission_ids:
        items:
          type: integer
        type: array
    type: object
...
  github_com_yamovo_contentx_internal_services.UpdateUserRequest:
    properties:
      avatar:
        type: string
      bio:
        type: string
      display_name:
        type: string
      email:
        type: string
      role_id:
        type: integer
      status:
        type: string
      website:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_handlers.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      err_code:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/internal_handlers.Meta'
    type: object
  internal_handlers.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: High-performance Headless CMS - API-first content platform
  title: ContentX API
  version: "1.0"
paths:
...
  /roles:
    get:
      description: Returns all roles (requires roles.view permission)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Role'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Creates a new role (requires roles.manage permission)
      parameters:
      - description: Role data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_services.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - Roles
  /roles/{id}:
    delete:
      description: Removes a role (requires roles.manage permission)
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Updates a role (requires roles.manage permission)
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_services.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Roles
  /roles/permissions:
    get:
      description: Returns all available permissions grouped by module
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: List permissions
      tags:
      - Roles
...
  /users:
    get:
      description: Returns users with pagination and filters (requires users.view
        permission)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Filter by role
        in: query
        name: role
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Search by name/email
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user (requires users.create permission)
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_services.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_services.SafeUser'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Soft-deletes a user (requires users.delete permission)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      description: Returns a single user by ID (requires users.view permission)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_services.SafeUser'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates a user (requires users.edit permission)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_services.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_services.SafeUser'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
  /users/{id}/reset-password:
    post:
      consumes:
      - application/json
      description: Resets a user's password (requires users.edit permission)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload {new_password}
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Reset user password
      tags:
      - Users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"